Zoo Animal Inventory README.md
This is a program about taking inventory of zoo animals. It will ask how many of each animal there are. Included animals are zebras, flamingos, tigers, lions, peacocks, capuchin monkeys and macaws. This program will not only count how many of each type of animal exists at the zoo, but also the total number of animals at the zoo.